requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
beautifulsoup4
google-api-python-client
requests
pandas
nltk
pyyaml
python-dotenv
pytest
jinja2
python-multipart
websockets
shodan
Pillow

# AI/ML Dependencies
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
